Improved documentation.
