+++ Pod/GearKit/Common/Extensions/CGFloatExtension.swift
@@ -23,7 +23,7 @@
 import Foundation
 import UIKit
 
-// Extension for CGFloat
+/// Extension for Swift Standard CGFloat type
 extension CGFloat {
     
     /// Check whether self is equal to another float provided with an epsilon (margin of error0
@@ -42,10 +42,9 @@ extension CGFloat {
     /// - parameter degrees: The degree value to convert to radians.
     ///
     /// - returns The value converted to radians.
-    static func degreesToRadians(degrees degrees: CGFloat) -> CGFloat {
+    public static func degreesToRadians(degrees degrees: CGFloat) -> CGFloat {
         
         let b = CGFloat(M_PI) * (degrees/180)
         return b
     }
-
 }
