+++ TODO.md
@@ -1,14 +1,13 @@
 # TODO
-[] Cleanup graphs
-[] Add assertions when needed
-[] Back button color is funky
-[] Add tests
-[] Add logger
-[] Local server with communication (Json/REST)
-[] Add TextView with placeholder and Done button
-[] Memory leaks
 [] SwiftDoc
 [] Add documentation as well in the ReadMe.md
+[] Add tests
+[] Remove storyboard and use only xibs as per https://medium.com/ios-os-x-development/dependency-injection-in-view-controllers-9fd7d2c77e55#.979fzwc6r
+[] Local server with communication (Json/REST)
+[] Back button color is funky
+[] Cleanup RadarGraphView
+[] Add TextView with placeholder and Done button
+[] Memory leaks
 [] Global review
 [] Add subpods for all modules
 [] Add icons for sample project
